requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1

# Web Automation
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
httpx==0.25.2

# NLP & AI
spacy==3.7.2
# Nota: Instalar modelo con: python -m spacy download es_core_news_sm

# Task Queue
celery==5.3.4
celery[redis]==5.3.4

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.39.1

# Security
cryptography==41.0.7
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0

# Utilities
pendulum==2.1.2
click==8.1.7
rich==13.7.0
tabulate==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-timeout==2.2.0
pytest-mock==3.12.0
faker==20.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
ipython==8.18.1

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# Image Processing (opcional)
pillow==10.1.0

# Async Support
aiofiles==23.2.1
aiohttp==3.9.1
asyncpg==0.29.0

# API Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
